tfpt-suite/requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
mpmath>=1.2.1
sympy>=1.10

# CMB Boltzmann solver (used by boltzmann_transfer)
camb>=1.6.0

# Plotting (modules may disable plotting via CLI)
matplotlib>=3.5.0

# Reporting (central PDF report)
reportlab
pillow

# Optional (not required for core modules)
# networkx>=2.8
# pandas>=1.3.0

